+++ src/Nethermind/Nethermind.AuRa.Test/Reward/AuRaRewardCalculatorTests.cs
@@ -95,7 +95,7 @@ namespace Nethermind.AuRa.Test.Reward
         [TestCase(9, 200ul)]
         public void calculates_rewards_correctly_before_contract_transition(long blockNumber, ulong expectedReward)
         {
-            _block.Header.Number = blockNumber;
+            _block.Header.Number = (ulong)blockNumber;
             AuRaRewardCalculator calculator = new(_auraParameters, _abiEncoder, _transactionProcessor);
             BlockReward[] result = calculator.CalculateRewards(_block);
             result.Should().BeEquivalentTo(new BlockReward(_block.Beneficiary, expectedReward));
@@ -114,7 +114,7 @@ namespace Nethermind.AuRa.Test.Reward
         [TestCase(15, 150ul)]
         public void calculates_rewards_correctly_after_contract_transition(long blockNumber, ulong expectedReward)
         {
-            _block.Header.Number = blockNumber;
+            _block.Header.Number = (ulong)blockNumber;
             BlockReward expected = new(_block.Beneficiary, expectedReward, BlockRewardType.External);
             SetupBlockRewards(new Dictionary<Address, BlockReward[]>() { { _address10, new[] { expected } } });
             AuRaRewardCalculator calculator = new(_auraParameters, _abiEncoder, _transactionProcessor);
@@ -140,7 +140,7 @@ namespace Nethermind.AuRa.Test.Reward
                 {50, _address50},
                 {150, _address150}
             };
-            _block.Header.Number = blockNumber;
+            _block.Header.Number = (ulong)blockNumber;
             BlockReward expected = new(_block.Beneficiary, expectedReward, BlockRewardType.External);
             SetupBlockRewards(new Dictionary<Address, BlockReward[]>() { { address, new[] { expected } } });
             AuRaRewardCalculator calculator = new(_auraParameters, _abiEncoder, _transactionProcessor);
@@ -152,7 +152,7 @@ namespace Nethermind.AuRa.Test.Reward
         [TestCase(15, 150ul)]
         public void calculates_rewards_correctly_for_uncles(long blockNumber, ulong expectedReward)
         {
-            _block.Header.Number = blockNumber;
+            _block.Header.Number = (ulong)blockNumber;
             _block = _block.WithReplacedBody(new BlockBody(_block.Body.Transactions, new[]
             {
                  Build.A.BlockHeader.WithBeneficiary(TestItem.AddressB).WithNumber(blockNumber - 1).TestObject,
