+++ src/Nethermind/Nethermind.Clique.Test/CliqueBlockProducerTests.cs
@@ -207,7 +207,7 @@ public class CliqueBlockProducerTests
             extraDataHex += "0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000";
 
             byte[] extraData = Bytes.FromHexString(extraDataHex);
-            BlockHeader header = new(parentHash, unclesHash, beneficiary, difficulty, number, gasLimit, timestamp, extraData);
+            BlockHeader header = new(parentHash, unclesHash, beneficiary, difficulty, (ulong)number, gasLimit, timestamp, extraData);
             Block genesis = new(header);
             genesis.Header.Hash = genesis.Header.CalculateHash();
             genesis.Header.StateRoot = new Hash256("0xba946bf2140ef68f7d9d57ef06a8ac0b28002b62060c462ba398389c97f1f1fa");
@@ -344,7 +344,7 @@ public class CliqueBlockProducerTests
         {
             if (_logger.IsInfo) _logger.Info($"ASSERTING HEAD BLOCK TIMESTAMP ON {nodeKey.Address}");
             Assert.That(
-                _blockTrees[nodeKey].FindBlock(_blockTrees[nodeKey].Head.Number - 1, BlockTreeLookupOptions.None).Timestamp + _cliqueConfig.BlockPeriod,
+                _blockTrees[nodeKey].FindBlock((long)(_blockTrees[nodeKey].Head.Number - 1), BlockTreeLookupOptions.None).Timestamp + _cliqueConfig.BlockPeriod,
                 Is.LessThan(_blockTrees[nodeKey].Head.Timestamp + 1));
             return this;
         }
@@ -363,7 +363,7 @@ public class CliqueBlockProducerTests
             WaitForNumber(nodeKey, number);
             if (_logger.IsInfo) _logger.Info($"ASSERTING {count} SIGNERS AT BLOCK {number}");
             BlockHeader header = _blockTrees[nodeKey].FindBlock(number, BlockTreeLookupOptions.None).Header;
-            Assert.That(_snapshotManager[nodeKey].GetOrCreateSnapshot(header.Number, header.Hash).Signers.Count, Is.EqualTo(count), nodeKey + " signers count");
+            Assert.That(_snapshotManager[nodeKey].GetOrCreateSnapshot((long)header.Number, header.Hash).Signers.Count, Is.EqualTo(count), nodeKey + " signers count");
             return this;
         }
 
@@ -373,7 +373,7 @@ public class CliqueBlockProducerTests
             WaitForNumber(nodeKey, number);
             if (_logger.IsInfo) _logger.Info($"ASSERTING EMPTY TALLY FOR {privateKeyB.Address} EMPTY AT {number}");
             BlockHeader header = _blockTrees[nodeKey].FindBlock(number, BlockTreeLookupOptions.None).Header;
-            Assert.That(_snapshotManager[nodeKey].GetOrCreateSnapshot(header.Number, header.Hash).Tally.ContainsKey(privateKeyB.Address), Is.EqualTo(false), nodeKey + " tally empty");
+            Assert.That(_snapshotManager[nodeKey].GetOrCreateSnapshot((long)header.Number, header.Hash).Tally.ContainsKey(privateKeyB.Address), Is.EqualTo(false), nodeKey + " tally empty");
             return this;
         }
 
@@ -401,7 +401,7 @@ public class CliqueBlockProducerTests
             while (Stopwatch.GetElapsedTime(startTime).TotalMilliseconds < _timeout)
             {
                 spinWait.SpinOnce();
-                if (_blockTrees[nodeKey].Head.Number >= number)
+                if (_blockTrees[nodeKey].Head.Number >= (ulong)number)
                 {
                     break;
                 }
